Move the `useDisplayTaxRate` hook to the `pages/settings/setup-tax-rate` directory.
