+++ js/src/pages/settings/setup-tax-rate/useDisplayTaxRate.test.js
